Change userpagevisit endpoint to a viewset.
List view now only returns objects you have permission to. You can
filter on user_id if you are staff. By changing to a viewset, but not
using the standard /endpoint/pk URL for updating (because we want
to update knowing only the user_id and page_id), we are determining
in the create method (POST on /endpoint/) whether we should create or
update.

Also included is a fix naming our CSRF cookie so it doesn't collide with
other IDAS and a fix for local logging of tests.