test_search_across_multiple_journals_with_AND_operator is flaky ====================================================================
The test_search_across_multiple_journals_with_AND_operator test in the
journals IDA is flaky. It was failing on multiple PRs and a rerun fixed
it.

AND was overridden to work as phrase search but test was testing it As
Logic. Now modified test according to "phrase" type search.